Move FooterNavBar into folder
